Update index.html
Made the links fancier. That's it. Literally.
It's a sunday night, ok? I should be sleeping but I'm procrastinating instead. So I accidentally open my own website and I'm like - yo, why does that say twitter? Why not my @? So I fix that. And I change blog to the blog's name. Because of reasons.

I'm tired.